+++ waves/tests/test_cartesian_product.py
@@ -1,20 +1,21 @@
 """Test CartesianProduct Class
 """
 
+import pathlib
 import pytest
 
-from waves.parameter_study import CartesianProduct
+from waves.parameter_generators import CartesianProduct
 
 class TestCartesianProduct:
     """Class for testing CartesianProduct parameter study generator class"""
 
     generate_io = {
-        'one_parameter': ({'parameter_1': [1, 2]}, ['set(parameter_1 "1")\n', 'set(parameter_1 "2")\n']),
+        'one_parameter': ({'parameter_1': [1, 2]}, ['parameter_1: 1\n', 'parameter_1: 2\n']),
         'two_parameter': ({'parameter_1': [1, 2], 'parameter_2': ['a', 'b']},
-                          ['set(parameter_1 "1")\nset(parameter_2 "a")\n',
-                           'set(parameter_1 "1")\nset(parameter_2 "b")\n',
-                           'set(parameter_1 "2")\nset(parameter_2 "a")\n',
-                           'set(parameter_1 "2")\nset(parameter_2 "b")\n'])
+                          ['parameter_1: 1\nparameter_2: a\n',
+                           'parameter_1: 1\nparameter_2: b\n',
+                           'parameter_1: 2\nparameter_2: a\n',
+                           'parameter_1: 2\nparameter_2: b\n'])
     }
 
     @pytest.mark.unittest
@@ -23,5 +24,6 @@ class TestCartesianProduct:
                              ids=generate_io.keys())
     def test_generate(self, parameter_schema, expected_text_list):
         TestGenerate = CartesianProduct(parameter_schema, None, False, False, False)
-        TestGenerate.generate()
+        parameter_set_file_paths = TestGenerate.generate()
         assert list(TestGenerate.parameter_study.values()) == expected_text_list
+        assert parameter_set_file_paths == [pathlib.Path(f"parameter_set{num}") for num in range(len(expected_text_list))]